Introduce batch clear for hot restart [ESC-9] (#2170)
**Reasoning:**
Collect hot-restart clear tasks and run them as a batch. This is to save
from shared queue space when there are high number of maps per
partition. If n is the map count, complexity will be O(1) instead of
O(n) for the occupied queue space.
As a bonus fsync call count will also be decreased in this manner.

**Fix:**
This is done by collecting keyPrefixes inside a long array and passing
it to clear-runnable.

GitOrigin-RevId: 19acece0d565d3b12443a7e7ab0869e1d59f3135