Use versioned serialization for all member to member communication [DEX-127] (#2308)
This PR uses versioned serialization for DS/IDS for all member-to-member
communication - it writes version bytes for all DS/IDS implementing
`@Versioned` interface (note: it *doesn't add* `@Versioned` on all
messages). Previously versioned serialization wasn't used in OS
member-to-member or in EE when RU wasn't licensed.

This change is required to allow the OS to EE restart, see #2163 for the
PR.

This PR doesn't change the rules around joining the cluster, so OS
members can't communicate with other members on different versions,
because they are not able to join the same cluster, similar to EE
members without the RU license.

Implementation:
- `EnterpriseDataSerializableSerializer` was moved to OS and renamed to
`VersionedDataSerializableSerializer`, it is used in `NodeExtension`
when a member is created (clients still use non-versioned
`DataSerializableSerializer`).
- `EnterpriseClusterVersionAware` was moved to OS and renamed to
`ClusterVersionAware`, OS uses an implementation reading the codebase
version, EE uses the same implementation as before.
GitOrigin-RevId: 7bd116b168dec57f6e220348b67efcce64ceec7e