Fix Rolling Upgrade 5.0 -> 5.3 [HZ-4431] (#2388)
In Platform 5.2 we introduced [improvements for
Kubernetes](https://github.com/hazelcast/hazelcast-mono/blob/master/docs/design/persistence/04-persistence-kubernetes-improvements.md).
For some reason, during the upgrade from 5.1 -> 5.2 cluster being
upgraded in `PASSIVE` state becomes `ACTIVE`, most probably, due to
logic in `KubernetesTopologyIntentTracker`. Moreover, despite the fact
all members were updated to 5.2, the **cluster version didn't change and
still is 5.1.**
During next upgrade 5.2 -> 5.3 inbound serialization stream holds
version 5.1 for `MembersUpdateOp`, there is a new field introduced in
`MemberInfo`, and the deserializer expects that `cpMemberUUID` is
present in `MemberInfo` list, which doesn't happen in this
circumstances.

This patch fixes a serialization exception[1] during continuous (5.0 ->
5.3) Rolling Upgrade, and actually unblocks it.
Still, RU is not perfect (with 5.2 members cluster still holds 5.1
version), but it is room for improvement in next cycles with more
dedicated focus on this issue, if it will be prioritized.

[1] - `HazelcastSerializationException: There is no suitable
de-serializer for type -509`

P.S, There are two JIRA issues for further investigation:
- https://hazelcast.atlassian.net/browse/ESC-33
- https://hazelcast.atlassian.net/browse/HZ-4809

GitOrigin-RevId: 64d6eb662d85865b52b7804799d4d691768c8dd0