Replace smart-routing with cluster-routing options [ADR-00002] (#2489)
This commit removes the existing `smart-routing` configuration option in
all usage other than loading from config, which is maintained for
backwards compatibility - this will intentionally remain for the
foreseeable future as this configuration option is very important and
low effort to maintain.

This option is replaced by the `cluster-routing` option, which has a
`mode` property which can be configured with `ALL_MEMBERS`,
`MULTI_MEMBER`, or `SINGLE_MEMBER`. These modes replace the old "smart",
"subset", and "unisocket" routing modes of the client.

Legacy support tests are included in `InvalidConfigurationClientTest`.

ADR: https://hazelcast.atlassian.net/wiki/x/AgBSKQE
GitOrigin-RevId: 8b6a912d4b16828e8e94bff3dcd727065ff12abe