Use METADATA group leader where applicable on CP client (#2559)
We can keep track of the current METADATA group ID when our client
receives updates, and then use that group ID to quickly retrieve the
current METADATA group leader when we want to retrieve a specific CP
group ID, or all CP group IDs.

We don't need to worry about setting the last known METADATA group ID to
null as an invalid group would return a `null` leader upon lookup, so we
can leave it to be updated on the next event received.

This PR also improves an existing flaky test in `CPGroupViewClientTest`
and introduces a new test to validate that proxy creation tasks are
executed on METADATA group leaders.

GitOrigin-RevId: d71d0a957f3cdcb7f9235c0a938f36444f9cda1e