Revert change in Registration [DEX-197] (#2587)
During the development of OS to EE upgrades
(https://github.com/hazelcast/hazelcast-mono/pull/2163) this change was
made to avoid a difference in serialization
- when doing RU from 5.2 to 5.5,
- OS to EE restart on 5.5, where in both cases the version was missing
(in 5.2 the class is not Versioned and in OS the serialization didn't
include the version at all.

However the design of OS to EE upgrades has changed during the
development - classes implementing Versioned are serialized with their
version also for OS members. Hence, this change is no longer needed.

Fixes https://github.com/hazelcast/hazelcast/issues/26366

GitOrigin-RevId: 6b4d9c74797518254d94a1bb2be08b231ea87ecc