Update Kafka libraries versions to those supported by Confluent [HZ-4815] (#2525)
[Confluent 7.6 only supports Kafka
3.6.x](https://docs.confluent.io/platform/current/installation/versions-interoperability.html#cp-and-apache-ak-compatibility),
but in hazelcast/hazelcast-mono#2498 ~we~ I upgraded to `3.7.1`.

[Kafka 3.7 introduced a breaking
change](https://github.com/apache/kafka/pull/14111), which [caused test
failures](hazelcast/hazelcast#26372).

As part of hazelcast/hazelcast-mono#2498, I also unified the version of
`kafka-clients` to use the same property (rather than just incrementing
the version to `3.7.1`) to ensure consistency going forwards.

I **have not reverted** this change, which means this PR will
effectively downgrade `kafka-clients` from `3.7.0` (the version _before_
the upgrade in hazelcast/hazelcast-mono#2498) to `3.6.1`.
The reason for this is that if Confluent 7.6 doesn't support Kafka 3.7,
there may be other issues like this our test suite hasn't exposed.

Fixes: hazelcast/hazelcast#26372 [[HZ-4815]]

[HZ-4815]:
https://hazelcast.atlassian.net/browse/HZ-4815?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: Burak Gök <burak.gok@hazelcast.com>
GitOrigin-RevId: ed25d207d11922a98ff4f9f00a0f2c609d330c3a