Resolve issues preventing Hazelcast v6 build [DI-181] (#2639)
Since upgrading [Hazelcast to
v6](https://github.com/hazelcast/hazelcast-mono/commit/31cfd36279d3e49965278e2e85bbb66efbe112cc),
build issues have been introduced.

There were four main issues - these are being merged simultaneously so
that we can have a ✅ in one build, as each issue is blocked by the
others:

### `attribution-maven-plugin`
Since incrementing Hazelcast to v6, the build now fails with errors
_like_:
```
[ERROR] Failed to execute goal on project hazelcast-enterprise-distribution: Could not resolve dependencies for project com.hazelcast:hazelcast-enterprise-distribution:pom:6.0.0-SNAPSHOT
[ERROR] dependency: com.hazelcast:hazelcast-enterprise:jar:6.0.0-SNAPSHOT (compile)
```

This was caused by `attribution-maven-plugin` executions trying to walk
the dependency tree, before the project's dependencies had actually been
compiled:
- `aggregated-attribution-enterprise`
- `aggregated-attribution-oss`

Because this dependency resolution behaviour is specified as an
annotation on the goal itself, irrespective of a `skip` parameter, it's
still triggering this resolution.
To resolve this, moved the declaration of the plugin from the "base" pom
into release-specific profile(s) - which does introduce duplication.

This also skips the generation of `aggregated-attribution.txt`, so the
the downstream consumer Maven assembly configuration tweaked to support
this file's presence being optional, by putting the file in a
subdirectory and then including the contents of the subdirectory
(whatever they may be) rather than explicitly referencing a file. **I've
not been able to test this** as even on `master` (i.e. without my
changes) I've not been able to get this goal to execute when building a
`distribution`.

### `hazelcast-enterprise-cdc-debezium`
`hazelcast-enterprise-cdc-debezium` is a dependency of the
`hazelcast-enterprise-distribution` module, but like the other
`extensions`, does not build out of the box. I've removed this
dependency to match how the OS `distribution` works.

### `com.hazelcast.internal.cluster.Versions#V6_0`
A constant for the new cluster version was missing.
This was originally raised
[here](https://github.com/hazelcast/hazelcast-mono/pull/2638).

### Test failures
Some tests are failing due to v6 being a new, major release. These are
**temporarily** `@Ignored`, to be addressed later:
-
`com.hazelcast.internal.rest.controller.ClusterActionTest.testChangeClusterVersion_Success`
- [HZ-4828](https://hazelcast.atlassian.net/browse/HZ-4828)
-
`com.hazelcast.jet.enterprise.JobCancelledAfterUpgradeTest.when_jobSubmittedToPreviousVersion_then_cancelledAfterUpgradeToCurrentVersion`
- [HZ-4828](https://hazelcast.atlassian.net/browse/HZ-4828)
-
`com.hazelcast.config.XmlConfigSchemaLocationTest.testSchemaLocationsExist`
- https://github.com/hazelcast/hazelcast-mono/pull/2646

Fixes: [DI-181](https://hazelcast.atlassian.net/browse/DI-181)

[DI-181]:
https://hazelcast.atlassian.net/browse/DI-181?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

[HZ-4828]:
https://hazelcast.atlassian.net/browse/HZ-4828?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

Co-authored-by: Tomasz Gawęda <tomasz.gaweda@outlook.com>
GitOrigin-RevId: 46dc991759dc1a98f2fc640513bf1c82dc4371be