Remove test dependency to Guava from Elasticsearch connector [HZ-4829] (#2692)
Guava is not needed any more in elastic tests since its usage there
could be replaced by `Map.of()` (this was not available in JDK 8 - in
the version of JDK for the tests were be prepared). This also fix
mentioned below test failure due to missing Guava in elastic integration
test.

Fixes https://github.com/hazelcast/hazelcast-mono/issues/2676

GitOrigin-RevId: 15b67ee24d2c62230ce655f908f127f84da66c38