Fix stale connections in niche MULTI_MEMBER routing [CORE-150] (#2678)
When we receive a new group view which does not include a member we are
already connected to (but does contain our primary member still), then
this connection becomes "stale" as it is not cleaned up. This scenario
is most likely encountered during a split-brain scenario, and although
it does self-correct after split-brain healing, the behaviour during a
split-brain is not correct when a partition group is split internally,
but not in relation to the client.

This PR addresses this by closing stale connections when receiving a new
members view, being careful not to close connections with invocations in
flight.

This PR also adds various useful `FINEST` logging entries, to aid in
debugging, and fixes some docker test environment issues for us poor
Windows users.

Fixes https://hazelcast.atlassian.net/browse/CORE-150
Closes https://github.com/hazelcast/hazelcast-mono/issues/2578

GitOrigin-RevId: 2c268d0fea56774ce27f917581fbfba4e603c656