Improve `checkstyle` performance by avoiding generating HTML report (#2370)
We use the [`checkstyle`
goal](https://maven.apache.org/plugins/maven-checkstyle-plugin/checkstyle-mojo.html)
of the `maven-checkstyle-plugin`:
> A reporting task that performs Checkstyle analysis and generates an
HTML report on any violations that Checkstyle finds.

We only use checkstyle for validation, not reporting, so instead we can
use the [`check`
goal](https://maven.apache.org/plugins/maven-checkstyle-plugin/check-mojo.html):
> Performs Checkstyle analysis and outputs violations or a count of
violations to the console, potentially failing the build.

This still fails the build for violations, but avoids creating an HTML
report, which makes slightly faster (~5%).

I also had to update lots of checkstyle paths to be up/down a directory
- I am unsure why this is required.

Benchmarked with:
```
./mvnw --file hazelcast/hazelcast/pom.xml -Dbuildtime.output.log=true clean checkstyle:{GOAL}
```

| Goal         	| Execution Time 	|
|--------------	|----------------	|
| `checkstyle` 	| 39.2s        	|
| `check`      	| 37.2s         	|

GitOrigin-RevId: 1fbe316273ccdadc5b3f2d2d7f682840394aee68