Fix `distribution.zip`'s `licenses/attribution.txt` generation (#2702)
In https://github.com/hazelcast/hazelcast-mono/pull/2639, the mechanism
by which `licenses/attribution.txt` was bundled into the
`distribution.zip` was changed.

Due to time constraints to unblock the build pipeline, this was not
properly tested or PR feedback actioned - this PR addresses that.

Changes:
- fixes the destination of the generated `attribution.txt` to ensure
it's picked up by the downstream assembly plugin
- adds a test to the existing license check suite to assert the file
exists in the `distribution.zip`s
- extracts the test from inside a GitHub action to a standalone shell
script
   - better IDE support
   - allows same script to be used for Jenkins job
- script refactored
   - improve readability
   - use `mvnw` over `mvn`
   - use pipes over extracting ZIP to disk
   - use `grep` exit codes over counting lines with `wc`
- puts back `nexus-staging-maven-plugin` which was removed accidentally

Co-authored-by: Josef Cacek <josef.cacek@gmail.com>
GitOrigin-RevId: 8fd11b46833d394df7ffce6e8d9ec4138794aeb9