Remove deprecated `goals` tag from `mongodb` `pom.xml` (#2698)
`mvnd clean` fails due to:
>[ERROR] Malformed POM
/hazelcast-mono/hazelcast/extensions/mongodb/pom.xml: Unable to read
model: Unrecognised tag: '{http://maven.apache.org/POM/4.0.0}goals' @
/hazelcast-mono/hazelcast/extensions/mongodb/pom.xml: ParseError at
[row,col]:[111,17]

The `plugin` has an `analyze` `goal` declared on the plugin itself, but
_also_ on it's `check-for-transitive-dependencies-usage` `execution`.
This was introduced in
https://github.com/hazelcast/hazelcast/pull/25744.

The [Maven spec](https://maven.apache.org/xsd/maven-4.0.0.xsd) declares
having a `goal` on the `plugin` itself as:
> **Deprecated**. Unused by Maven

As such, removed.

NB - [this is a bug in
`mvnd`](https://github.com/apache/maven-mvnd/issues/859), but a cleanup
in our codebase works around it.

GitOrigin-RevId: 4d7be0b2231e340f2081d2ac24aa88b8c96fa554