Fix flaky IOBalancerMemoryLeakTest [HZG-54] (#2869)
This test was failing because the `HttpClient` used for sending REST
endpoint requests was holding on to connections. The `HttpClient`
documentation states that connections are cleaned up as the client
object is GC'd, and so this test can be resolved by simply invoking a GC
using `System.gc()` after all of our REST requests have been submitted.

Fixes https://hazelcast.atlassian.net/browse/HZG-54
Closes https://github.com/hazelcast/hazelcast/issues/9216

GitOrigin-RevId: d34026794626a68e72ee610d6cdbc172dd9f6952