Return error if attempting to shutdown cluster before Join complete [ESC-22] (#2855)
If cluster shutdown is attempted before the node has joined the cluster
a NPE will occur as the master address is not set. In this case it would
appear to the user that the request was successful, but the cluster will
still be alive.

In this case we return a CONFLICT HTTP error and don't attempt any
shutdown.

Checklist:
- [X] Labels (`Team:`, `Type:`, `Source:`, `Module:`) and Milestone set
- [X] Add `Add to Release Notes` label if changes should be mentioned in
release notes or `Not Release Notes content` if changes are not relevant
for release notes
- [X] Request reviewers if possible
GitOrigin-RevId: eda08d60684812ffa2ccf506e683ab91cb9a610e