Fix the NPE in the JetServiceBackend.shutdown() method when the member fails fast [DEX-156] (#2899)
Fix the NPE in the JetServiceBackend.shutdown() method when the member
fails fast due to exception in the REST service constructor.
The NPE occurs when JetServiceBackend's shutdown method is called before
it is initialized. This happens because an error during the
initialization of the REST service causes a fast-fail, triggering the
shutdown of all services. At that moment because JetServiceBackend's
init method has not been called yet, its shutdown method invoked without
proper initialization.

issue: https://github.com/hazelcast/hazelcast-mono/issues/2217
related comment:
https://github.com/hazelcast/hazelcast-mono/issues/2217#issuecomment-2245998062
GitOrigin-RevId: bdfd8a9c8f47692d8e8722a9097db49a90854893