Fix TTL being ignored on load from map store [ESC-36] (#2841)
Issue raised from customer shows that after data has been expired from a
map (with a TTL) and reloaded from a `MapStore` upon calling `map.get`
the `MapConfig` TTL is ignored on the reload and instead, an expiry time
of `Long.MAX_VALUE` is applied to the record.

Note - If the reloaded data is updated using `map.put` on existing keys,
the TTL does takes affect.

Opted to store the actual TTL or -1 in the BiTuple.
Introduced tests from the reproducer.
Added missing test coverage.

Cause of regression :
https://github.com/hazelcast/hazelcast-mono/commit/f70425e02691ebd5d9484c5e33f3d13a10ff7349

Checklist:
- [X] Labels (`Team:`, `Type:`, `Source:`, `Module:`) and Milestone set
- [X] Add `Add to Release Notes` label if changes should be mentioned in
release notes or `Not Release Notes content` if changes are not relevant
for release notes
- [X] Request reviewers if possible
GitOrigin-RevId: fce5127b4e435090327e9e0ec2ee370a7b2677f4