Simplify ClientHazelcastRunningInForkJoinPoolTest [HZG-55] (#2875)
The test altered in this PR frequently fails Jenkins builds across
multiple JVMs and operating systems.

The test itself seems overly complicated for what it’s trying to
actually test: the goal of the test is to validate that when you have
slow data loading for 1 client's map, it does not block an entry
listener addition for another map on the server. We can simplify the
test to validate this behaviour without involving checks that are likely
to flake in our Jenkins environment.

It appears that the ForkJoinPool is not actually relevant, despite being
the basis of the test class name (the server listener addition does not
use the FJP), but I have kept the client using it anyway.

Fixes https://hazelcast.atlassian.net/browse/HZG-55
Closes https://github.com/hazelcast/hazelcast/issues/24850

GitOrigin-RevId: 631f1b63f7d43ccdcd0af72f462df69fc3896b99