Fix some invocations not de-registered due to invocation method [HZG-155] (#3046)
This PR falls into 3 main parts, resolving the issue identified in
HZG-155:

1. `OperationService` improvements
2. Incorrect `invokeOnTarget` call updates
3. Introducing `SelfResponseOperation` marker interface

**OperationService improvements**

The first part of this PR focuses on improving the javadocs for
`OperationService` to warn about calling methods which return `Future`
instances with Operations that do not expect a response. This can cause
invocations to not be de-registered, leading to increasing memory
consumption.

It also adds a new method to `OperationService` and
`OperationServiceImpl` called `executeOrSend`, intended to fulfill the
same purpose as `invokeOnTarget`, but geared for Operations which do not
expect a response.

Assertions have been added to `OperationServiceImpl` methods that return
`Future` objects to aid in catching areas of the codebase that need
updating. The idea is that the Jenkins jobs on this PR will cause these
assertions to be triggered, flagging areas requiring improvement, which
will be addressed in this PR before merging.

**Incorrect invokeOnTarget call updates**

The second part of this PR is applying updates to areas of the codebase
that call `invokeOnTarget` with operations that do not expect a
response, correcting their behaviour to use the new `executeOrSend`
method instead.

**Introducing `SelfResponseOperation` marker interface**

The final part of this PR is introducing a new marker interface,
`SelfResponseOperation`, which is used to mark Operations which call
their own `sendResponse` method internally. This is then checked within
`OperationService` calls by an `assert`, preventing future issues of
Operations which don't use `sendResponse` being called by methods which
return `Future`s.

Fixes https://hazelcast.atlassian.net/browse/HZG-155

GitOrigin-RevId: 249a0db8602a7b22e55aa402d2465c978198c2b0