Introduce phone home for jet connectors [DEX-261] (#3041)
Adds phone home collection and test for fileWatcher, fileBuilder and
Avro source&sink.

Phone home service is not implemented yet, the impact of merging this
without it is that a little bit more data will exist in the phone home
ping, but they'll be ignored in the phone home service backend.

This PR introduces a new interface ConnectorNameAware. The
ProcessorMetaSuppliers that are associated with a source or a sink will
implement this interface.

In the ExecutionPlanBuilder (which means in the master member that
prepares ProcessorSuppliers for a job), these PMS will be handled and in
the JetServiceBackend a ConcurrentHashMap will store the usage counts
for every connector. Every connector has a unique name in ConnectorNames
class. For extension connectors, they have a constant in their own
module.

Connector names follow the following structure:

"\<connectorIdentifier\>(Source|Sink)"  and follows camel case.

e.g avroSink, fileSource, fileWatcherSource, avroSource and so on.

---------

Co-authored-by: devopshazelcast <devops@hazelcast.com>
Co-authored-by: Nishaat Rajabali <12186256+nishaatr@users.noreply.github.com>
Co-authored-by: Jack Green <jack.green@hazelcast.com>
Co-authored-by: Tomasz Gawęda <tomasz.gaweda@outlook.com>
Co-authored-by: Ondrej Lukas <ondrej@hazelcast.com>
Co-authored-by: James Holgate <130981049+JamesHazelcast@users.noreply.github.com>
Co-authored-by: Burak Gök <burak.gok@hazelcast.com>
GitOrigin-RevId: 9c4a74d355d60a39e460a70c0e38a68106b9eeac