Fix partition table snapshot selection [HZG-152] (#3064)
As of #19105[1], we use stored snapshots of the partition table for rejoining members if persistence is enabled. However, when determining the partition table candidates, it is checked whether every replica in the snapshot belongs to a data member. This causes snapshots that have no assignment for the rejoining member to be marked as candidates. Since such snapshots are less "distant" to the current partition table, they are selected. To resolve this issue, when determining candidates, it should instead be checked whether every data member owns a replica in
the snapshot.

Fixes https://hazelcast.atlassian.net/browse/HZG-152

[1]: https://github.com/hazelcast/hazelcast/pull/19105/files#diff-8a8c17fa07bc1df1c4bdba408f88c330f2ac3f5318c13b3a295d2e916e0be021R886-R889

GitOrigin-RevId: 0d0b492022a0fde564e355c2600d7cbeddc7c8b5