Support customizing NodeContext in TestHazelcastInstanceFactory [HZG-184] (#3197)
Lack of NodeContext customization in TestHazelcastInstanceFactory has led people to use HazelcastInstanceFactory directly. To prevent this, DelegatingNodeContext is introduced, which by default delegates all operations to the default NodeContext (DefaultNodeContext for real network, MockNodeContext for mock network) and allow overriding selected behavior. As an example, a delegating variant of FirewallingNodeContext is introduced, which wraps the Server provided by the delegate in a FirewallingServer.

GitOrigin-RevId: 7b262232987cc72b054793fdb7bdbe8e7072d803