Support customizing NodeExtension in TestHazelcastInstanceFactory [HZG-184] (#3201)
In https://github.com/hazelcast/hazelcast-mono/pull/3197, NodeContext customization was introduced to TestHazelcastInstanceFactory to avoid direct usage of HazelcastInstanceFactory. In our tests, this need is also addressed by subclassing TestHazelcastInstanceFactory. All such usages can now be rewritten to utilize DelegatingNodeContext, which was introduced in #3197. However, when looking at the usages, it is seen that
NodeExtension is customized predominantly. That's why we introduce NodeExtensionCustomizer, a variant of DelegatingNodeContext that allows overriding NodeExtension, and a convenience method to TestHazelcastInstanceFactory to facilitate its usage.

GitOrigin-RevId: 78982903a1857fe1f56a7861ffe7c12c704ab391