Introduce new method with maxBatchSize in PythonTransforms [SUP-497]  (#3206)
When using `PythonTransforms`, the whole inbox for this stage is passed
in a single batch to the backed service. If the Python service is slow
(e.g. computing vector embeddings) and a backlog builds up, the size of
the message passed to the backend will eventually exceed GRPC maximum
message size limit.

This PR allows the user to use a smaller batch size by providing
additional methods `mapUsingPython` and `mapUsingPythonBatch` with
`maxBatchSize` parameter.

Co-authored-by: Frantisek Hartman <frant.hartm@gmail.com>
GitOrigin-RevId: d2ee21ad53ecbd7d0772de3a70825b0bb27edab0