K8s shutdown issues [HZG-110] (#3204)
Root cause is a data race occurring on graceful shutdown.

After first member is removed from the cluster a
`SafeStateCheckOperation` is invoked concurrently on different nodes. In
some cases this invocation is executed moments before the member is
removed from `MembershipManager` and so the operation is put into a
retry loop, using a stale view of the member list.

This causes the invocation to retry on a member that has already
shutdown and so the operation times out and the trying member does not
execute subsequent code paths for graceful shutdown, including the
`HotRestartPersistenceEngine`.

This causes `chunk.active` to not get flushed and closed, so data is
lost.

See
[here](https://hazelcast.atlassian.net/browse/HZG-110?focusedCommentId=102655)
for deeper explanation.

May also resolve HZG-154, requires further testing

Checklist:
- [X] Labels (`Team:`, `Type:`, `Source:`, `Module:`) and Milestone set
- [X] Request reviewers if possible
GitOrigin-RevId: 1cdbff4eb963b227b48cb970de6af1c0f409b9d6