Fix NodeEngine context missing in cached executor service [HZG-186] (#3205)
This PR also removes the redundant `NodeEngine` parameter being passed
to `CachedExecutorServiceDelegate`, which is not needed as the
underlying `ExecutorService` handles this context.

Introducing a regression test for this scenario is difficult, as
currently testing approaches (such as
`CachedExecutorServiceDelegateTest`) provide a concrete `ThreadFactory`,
which involves explicitly providing (or not) a `NodeEngine` parameter,
meaning we would not be testing anything meaningful. This is further
complicated by the customer's reproducer involving Spring, and a lack of
Spring testing for UCN in general.

To try and make up for this lack of test coverage, I reviewed all
constructions of `PoolExecutorThreadFactory` to ensure there are no
other member-side gaps present. I also updated the javadocs for
`PoolExecutorThreadFactory` to ensure the correct constructor is used
for client or member contexts in the future.

Fixes https://hazelcast.atlassian.net/browse/HZG-186

GitOrigin-RevId: 97e69ada8328dbf095d1ac82bb0244aaf2d8c058