Add control of search list size in similarity search [AI-140] (#3229)
The most important parameter that affects search results quality is
search list size. Before this change it was always equal to `topK` or
`partitionLimit` (if supplied), which produced poorer quality results
with smaller values.

Explicit control of that parameter is introduced via `efSearch` hint.
This should improve in particular searches with small topK (e.g. 10),
where in some benchmarks it was hard to get >95% precision at all.

`partitionLimit` hint remains unchanged and can be used for further
performance tuning. During initial tests it was observed that with good
`efSearch` value, `partitionLimit` can be smaller than `topK` without
compromising the quality.

Benchmarking was performed using `dbpedia-openai-1M-1536` dataset which
has default topK = 10 on Apple M1 laptop, JDK 21 and Panama enabled.
Single-member cluster with `-Xmx10g` and 16 partitions was used. Index
parameters were max_degree = 16 and 32 (this is after upgrade to JVector
3, equivalent to 8 and 16 respectively in previous benchmarks),
ef_construction = 128 and disabled deduplication.

![image](https://github.com/user-attachments/assets/462d2502-5865-4301-a063-6a4ad61c77a1)

![image](https://github.com/user-attachments/assets/040f2121-730a-4be1-9b2b-f5093a3651df)

Labels on the line are `efSearch` value.

The results show that precision depends predominantly on `efSearch`.
`partitionLimit` has very small impact when `efSearch` is sufficiently
large. For reference, results with topK=100 are provided (previously
recommended way to improve quality in low-topK searches). Quality is
similar for the same effective search list size, but searches with
inflated `topK` are slower.

Increasing `efSearch` past 100 brings very little improvement in quality
but significant increase in run time (in this case).

Important to note about these benchmarks:

1. the results are quite noisy, likely due to running on a laptop, but
they prove the main point about precision. Proper benchmarks will be
done later.
2. search with `partitionLimit=10` is faster (!) than with
`partitionLimit=5`. This however may be an artifact of benchmark
methodology, lack of warmup, JIT compilation. This will be investigated
further.
3. from initial tests is seems that `efSearch` value around 50-100
should be a good choice, but the sample is very small (only a few tests)
so this may not turn out to be true.
4. `partitionLimit=3` results in slightly but visibly decreased quality,
even though 3*16=48 entries in total are fetched from partitions for
topK=10 query. It seems that the probability of omitting good candidate
(because it is 4th best in partition) is non-negligible in this case.
GitOrigin-RevId: a9e809c711d825f81a3f6b548ec0ee95c68a37d6