Fix HazelcastProxyFactoryTest & MavenInterface [HZG-195] (#3339)
There is currently a new constant test failure related to
`HazelcastProxyFactoryTest#testProxyHazelcastInstanceClasses_ofSameVersion_areSame()`.
The test is failing after [this
commit](https://github.com/hazelcast/hazelcast-mono/commit/d16ef527cad4d4ecfce9df8c60c8872a854b5f5f)
due to the removed `destroy()` methods in `TestSerializerHook`.

This `destroy()` method was made a `default` method of the interface [in
2020](https://github.com/hazelcast/hazelcast-mono/commit/088f3680e6571e7091b6c4cd12cdeaaedbeefba1),
but the `HazelcastProxyFactoryTest` uses HZ version `4.0`, which is
before this change was made, and so it complains that this method is not
implemented by `TestSerializerHook`.

There are two options to resolve here - we could reinstate the
`destroy()` method within `TestSerializerHook`, or we could update the
test to use a newer version of Hazelcast - given that `4.0` is well out
of scope for support or maintenance at this point, I have opted for the
latter in this PR.

This PR also fixes some issues with `MavenInterface` (as discovered in
local testing) by introducing the `--quiet` flag to `MavenInterface`'s
expression evaluation to only return the result, and omit the
surrounding fluff (this was breaking resolution for me and MacOS users).

During this work I did discover some issues with the `mvnw.cmd` file as
it does not handle spaces in usernames correctly. To reduce the
complexity of this fix PR, it is not included here and will be addressed
separately.

Fixes https://hazelcast.atlassian.net/browse/HZG-195
Closes https://github.com/hazelcast/hazelcast/issues/26433

GitOrigin-RevId: 3ee61102aaf384f9c84074582f7c14b35286df57