Do not destroy UCN NodeEngine context in OperationExecutorImpl [HZG-193] (#3453)
It is safe to remove this cleanup as it will be performed in
`DefaultNodeExtension#onThreadStop` anyway. The only way the
`NodeEngine` can be invalidated is by restarting the JVM.

In the in `OperationExecutorImpl#shutdown` method we are explicitly
waiting for some threads to shutdown (not terminating), so we shouldn't
clear context at this stage.

This issue was discovered during development of
https://github.com/hazelcast/hazelcast-mono/pull/3386 which terminates
an instance on it's own operation thread to simulate a coordinator crash
during Jet snapshot operations.

Checklist:
- [x] Labels (`Team:`, `Type:`, `Source:`, `Module:`) and Milestone set
- [x] Add `Add to Release Notes` label if changes should be mentioned in
release notes or `Not Release Notes content` if changes are not relevant
for release notes
- [x] Architecture Design Record (ADR) reviewed and signed-off if this
PR represents a significant architectural change
- [x] Request reviewers if possible
- [x] New public APIs have `@Nonnull/@Nullable` annotations
- [x] New public APIs have `@since` tags in Javadoc
- [ ] Send backports/forwardports if fix needs to be applied to
past/future releases
GitOrigin-RevId: 6ab998c52b5b94d657b5d62590cb842e389089fb