Add UCN support for Jet jobs [HZG-88] (#3293)
This commit adds support for User Code Namespaces in Jet Jobs, defined
as a `JobConfig` argument. This functionality works independently of
existing Jet custom resource support so as not to impact backwards
compatibility support. The defined User Code Namespace will be used as
the parent classloader for the job, maintaining all other logic for
existing Jet job classloading.

Support for UCN, which provides classloaders which persist beyond the
Jet job lifecycle, will fix various known issues with Jet job resources.
The use of a consistent `ClassLoader` will mean jobs can share classes
with other user code structures (such as domain objects in `IMap`), and
not encounter exceptions caused by a Jet-specific `ClassLoader` class
clashing with a class defined by another `ClassLoader`. Particularly
when using a static UCN which is not dynamically updated, this will
allow Jet jobs to be undeployed and redeployed, and still be able to
interact with data created by the previous job iteration (previously
this would encounter the exception mentioned earlier).

Strictly speaking, this does not add any "new" custom resource
functionality to Jet, as it already supports resource uploading, but it
does improve Jet's compatibility between jobs using custom resources,
and also improves Jet's compatibility when handling custom classes used
in other data structure (such as a domain object in `IMap`). In the
future it may make sense to deprecate and eventually remove "resource
upload", in favour of making UCN the first class choice.

Due to the dynamic support nature of UCN, the introduction of UCN
support in Jet will introduce the possibility for small race scenarios
if a Jet job is submitted while a UCN is being cleaned up (it may use
the old UCN or new, indeterminate). This could also happen when new
members join the cluster with different UCN deployments, before they are
synchronized with the cluster. Caveats such as this will need to be well
documented, and advice to users will be for Jet jobs to only reference
UCNs which do not change at runtime.

This commit also adds some tests for UCN support in Jet on both embedded
instances and clients, as well as client security tests.

Closes https://hazelcast.atlassian.net/browse/HZG-88

GitOrigin-RevId: 7ed1ddddff0e9b1d89eed8b72d2a7ab23faaa8e3