CP: Enhance error message that the client receives when the CP system is reset [HZG-7] (#3430)
On the client side, since there is no lock on the cluster side after the
restart of the cluster, the NotLeaderException is thrown with a message
"null is not LEADER of null. Known leader is: N/A".
Because this message is not meaningful for the users, I introduced a new
custom message for NotLeaderException when FencedLock is no longer valid
for the cluster.
Added a new constructor to NotLeaderException for cases where the
cluster has no info about the client's lock.

Fixes [SUP-117](https://hazelcast.atlassian.net/browse/SUP-117)

[SUP-117]:
https://hazelcast.atlassian.net/browse/SUP-117?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

GitOrigin-RevId: 20bdb28d27f7dd3c21cc93de5e5ef365b40bc89a